updated stagzmerfolk to 1.5

+++ Source/StagzMerfolk/HarmonyPatches/Pawn_PathFollower_Patches.cs
@@ -14,7 +14,7 @@ namespace StagzMerfolk.HarmonyPatches;
 })]
 public static class Pawn_PathFollower_Patches
 {
-    private static void Postfix(Pawn pawn, IntVec3 c, ref int __result)
+    private static void Postfix(Pawn pawn, IntVec3 c, ref float __result)
     {
         if (pawn?.genes != null && pawn.genes.HasGene(StagzDefOf.Stagz_Aquatic) && pawn.Map.terrainGrid.TerrainAt(c).IsWater)
         {
